requirements.txt
accelerate==1.6.0
adapters==1.1.1
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
blobfile==3.0.0
cached-property==2.0.1
certifi==2025.1.31
charset-normalizer==3.4.1
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
docopt==0.6.2
executing==2.2.0
ffmpeg==1.4
filelock==3.18.0
fsspec==2025.3.2
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
llvmlite==0.44.0
lxml==5.3.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
more-itertools==10.6.0
mosestokenizer==1.2.1
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.4
ollama==0.4.7
openai-whisper==20240930
openfile==0.0.7
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.7
prompt_toolkit==3.0.51
protobuf==6.30.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycryptodomex==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
pydub==0.25.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-telegram-bot==22.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentencepiece==0.2.0
six==1.17.0
skops==0.11.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.13.1
tabulate==0.9.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toolwrapper==2.1.0
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uctools==1.3.0
urllib3==2.4.0
wcwidth==0.2.13
wtpsplit==2.1.5
